chore: We split existing services into binaries and crates
We split our services into binaries and crates to make it easy for sharing code.
