+++ Cargo.toml
@@ -44,6 +44,7 @@ needless-pass-by-ref-mut = "warn"
 string-lit-as-bytes = "warn"
 
 [workspace.dependencies]
+cadence = "1.4"
 clap = { version = "4.0", features = ["derive", "env"] }
 tokio-tungstenite = { version = "0.26", features = ["native-tls"] }
 futures-util = "0.3"
@@ -57,14 +58,15 @@ dirs = "6.0"
 arboard = "3.4"
 dotenvy = "0.15.7"
 async-trait = "0.1"
+mempool-rebroadcaster = { path = "./crates/mempool-rebroadcaster" }
+sidecrush = { path = "./crates/sidecrush" }
+metrics = "0.24.1"
+metrics-derive = "0.1"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", default-features = false, features = ["env-filter", "fmt", "ansi", "json"] }
 tokio = { version = "1.0", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-metrics = "0.24.1"
-metrics-derive = "0.1"
-cadence = "1.4"
 
 # alloy
 alloy-primitives = { version = "1.5.2", default-features = false, features = [
